requirements.txt
# ARES-GBM Research Tools — Dependencies

# Core metabolic modeling
cobrapy>=0.26.0

# Data analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Survival analysis
lifelines>=0.27.0

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0
kaleido>=0.2.1

# SBML / metabolic model I/O
python-libsbml>=5.20.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
